+++ benchopt/plotting/default_plots.py
@@ -51,7 +51,7 @@ class ObjectiveCurvePlot(BasePlot):
                 **self.get_style(solver)
             }
 
-            if X_axis == "Time":
+            if X_axis == "Time" and df_filtered['idx_rep'].nunique() > 1:
                 # Compute the quantiles for runtime
                 q = (
                     df_filtered.groupby('stop_val')['time'].quantile([.2, .8])
@@ -246,4 +246,6 @@ class TablePlot(BasePlot):
         return {
             "title": f"{objective}\nData: {dataset}",
             "columns": columns,
+            "default_order_ascending": True,
+            "default_order_column": 0,
         }